[riscv64]working on riscv64 platform

+++ src/arch/riscv64/include/types.h
@@ -41,11 +41,11 @@ typedef unsigned long long		physical_addr_t;
 typedef unsigned long long		physical_size_t;
 
 typedef struct {
-	volatile long counter;
+	volatile int counter;
 } atomic_t;
 
 typedef struct {
-	volatile long lock;
+	volatile int lock;
 } spinlock_t;
 
 #ifdef __cplusplus
